Code is failing when we are inserting to product image relation table. Having issues with Kafka Consumer thread
