+++ embedding_generator/urls.py
@@ -1,7 +1,7 @@
-from django.urls import path
-from embedding_generator.views import process
-
-
-urlpatterns = [
-    path('process/', process, name='process_response'),
-]
+# from django.urls import path
+# from embedding_generator.views import process
+#
+#
+# urlpatterns = [
+#     path('process/', process, name='process_response'),
+# ]
