+++ embedding_generator/views.py
@@ -1,9 +1,9 @@
-from django.http import JsonResponse
-from django.shortcuts import render
-from embedding_generator.task import process_images
-
-
-# Create your views here.
-def process(request):
-    process_images()
-    return JsonResponse({'status': 'images downloaded and embeddings created'}, status=200)
+# from django.http import JsonResponse
+# from django.shortcuts import render
+# from embedding_generator.task import process_images
+#
+#
+# # Create your views here.
+# def process(request):
+#     process_images()
+#     return JsonResponse({'status': 'images downloaded and embeddings created'}, status=200)
